Code: Added customizable paths for the sqlite database
Now the user may set a custom path for the database by setting $CORKDB.
If not set then the app will default to using XDG_DATA_HOME as defined
in it's standard.
If all else fails, then we just place the db in the directory from where
the program ran.

Also added $CORKDB_TEST , which when set places the database in our
running directory. The intention here is to not mess with the dev's own
database when running tests or running through cargo run.