Code: eased marking items as read
Simiplified marking items by switching from hashes to positions
stored on a table. These numbers change every time the user
checks their "new" (i.e. not marked) items, keeping them managable and
easy to write.

Added consistency between the other tables and this new quickmarks
table.

I belive that the mark system may be a bit over-thought and could be
simplified, but for now works. To improve it further the exact text we
are printing on screen needs to become clearer and easier to understand.