Fix: Hashes will now be stable after upgrades of corkboard
I was wrongly using the _DefaultHasher_ from the rust stdlib, that
explicitly states should not be relied upon over releases. The result
was that updating some feeds would fail due to collisions in the PK
of items table. IE the hasher would change and an old entry would clash
with a newer one.

I moved to sha256 provided by another package. This way we'll be sure
hashes stay stable over time.

Also added a sourceable script to ensure the safety of the dev's
personal database when doing development.