Feat: Atom support
Corkboard can now consume Atom feeds as if they were RSS feeds.
Note that Atom's spec is not implemented fully (and in the case of
the channel/feed link we are stricter since we do require it).

Wrote two sample atom feeds and an integration test _add-atom-test.rs_
for them.

Modified _xml_to_rss_ to decide which type of feed was fed to it before
moving on to parsing. (So that users don't need to specify it
themselves).